pyproject.toml
[tool.poetry]
name = "e6py"
version = "0.1.0"
description = "An e621 API wrapper"
authors = ["zevaryx <zevaryx@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^22.1.0"
requests = "^2.28.1"
sentinel = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
python-dotenv = "^0.21.0"
pytest-cov = "^3.0.0"

[tool.coverage.run]
source = ["e6py"]

[tool.pytest.ini_options]
addopts = "--cov --cov-report term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
